requirements.txt
click==8.1.2
dlib==19.23.1
face-recognition==1.3.0
face-recognition-models==0.3.0
importlib-metadata==4.11.3
imutils==0.5.4
numpy==1.21.5
opencv-python==4.5.5.64
picamera==1.13
Pillow==9.1.0
typing_extensions==4.1.1
zipp==3.8.0
